Extract runtime orchestration collaborators
