+++ src/backend/services/ratchet/service/ratchet-decision-logging.helpers.ts
@@ -1,0 +1,128 @@
+import type { createLogger } from '@/backend/services/logger.service';
+import type { RatchetState } from '@/shared/core';
+import type {
+  PRStateInfo,
+  RatchetAction,
+  RatchetDecisionContext,
+  WorkspaceWithPR,
+} from './ratchet.types';
+import {
+  buildFailedCheckDiagnostics,
+  buildReviewTimestampDiagnostics,
+  buildSnapshotDiagnostics,
+} from './ratchet-pr-state.helpers';
+
+type Logger = ReturnType<typeof createLogger>;
+
+function describeNonRatchetingReason(
+  action: Exclude<RatchetAction, { type: 'TRIGGERED_FIXER' }>
+): string {
+  switch (action.type) {
+    case 'WAITING':
+      return action.reason;
+    case 'FIXER_ACTIVE':
+      return `Ratchet fixer session is already active (${action.sessionId})`;
+    case 'DISABLED':
+      return action.reason;
+    case 'COMPLETED':
+      return 'PR is already merged';
+    case 'ERROR':
+      return action.error;
+  }
+  const exhaustiveCheck: never = action;
+  throw new Error(`Unhandled ratchet action: ${JSON.stringify(exhaustiveCheck)}`);
+}
+
+export function buildRatchetingLogContext(params: {
+  workspace: WorkspaceWithPR;
+  previousState: RatchetState;
+  newState: RatchetState;
+  action: RatchetAction;
+  prStateInfo: PRStateInfo | null;
+  prNumber: number | null;
+  decisionContext: RatchetDecisionContext | null;
+  logger: Logger;
+}) {
+  const {
+    workspace,
+    previousState,
+    newState,
+    action,
+    prStateInfo,
+    prNumber,
+    decisionContext,
+    logger,
+  } = params;
+  const reviewDiagnostics = buildReviewTimestampDiagnostics(
+    workspace,
+    prStateInfo,
+    decisionContext,
+    logger
+  );
+  const snapshotDiagnostics = buildSnapshotDiagnostics(workspace, prStateInfo, decisionContext);
+  const latestReviewActivityAt = reviewDiagnostics.latestReviewActivityAtMs;
+
+  return {
+    workspaceId: workspace.id,
+    prUrl: workspace.prUrl,
+    prNumber,
+    prState: prStateInfo?.prState ?? null,
+    ciStatus: prStateInfo?.ciStatus ?? null,
+    hasChangesRequested: prStateInfo?.hasChangesRequested ?? null,
+    hasMergeConflict: prStateInfo?.hasMergeConflict ?? null,
+    snapshotKey: prStateInfo?.snapshotKey ?? null,
+    ciSnapshotKey: snapshotDiagnostics.ciSnapshotKey,
+    snapshotComparison: snapshotDiagnostics.snapshotComparison,
+    previousState,
+    newState,
+    ratchetEnabled: workspace.ratchetEnabled,
+    ratchetActiveSessionId: workspace.ratchetActiveSessionId,
+    ratchetLastCiRunId: workspace.ratchetLastCiRunId,
+    ciStatusCheckRollup: prStateInfo?.statusCheckRollup ?? null,
+    ciFailedChecks: buildFailedCheckDiagnostics(prStateInfo),
+    prReviewLastCheckedAt: workspace.prReviewLastCheckedAt?.toISOString() ?? null,
+    latestReviewActivityAt: latestReviewActivityAt
+      ? new Date(latestReviewActivityAt).toISOString()
+      : null,
+    reviewTimestampComparison: reviewDiagnostics.reviewTimestampComparison,
+    actionType: action.type,
+  };
+}
+
+export function logWorkspaceRatchetingDecision(params: {
+  workspace: WorkspaceWithPR;
+  previousState: RatchetState;
+  newState: RatchetState;
+  action: RatchetAction;
+  prStateInfo: PRStateInfo | null;
+  decisionContext?: RatchetDecisionContext | null;
+  logger: Logger;
+}): void {
+  const { workspace, previousState, newState, action, prStateInfo, logger } = params;
+  const decisionContext = params.decisionContext ?? null;
+  const prNumber = prStateInfo?.prNumber ?? workspace.prNumber;
+  const prNumberLabel = prNumber ?? 'unknown';
+  const workspacePrPrefix = `workspace ${workspace.id} for PR #${prNumberLabel}`;
+  const context = buildRatchetingLogContext({
+    workspace,
+    previousState,
+    newState,
+    action,
+    prStateInfo,
+    prNumber,
+    decisionContext,
+    logger,
+  });
+
+  if (action.type === 'TRIGGERED_FIXER') {
+    logger.info(`Ratcheting ${workspacePrPrefix}`, {
+      ...context,
+      sessionId: action.sessionId,
+      promptSent: action.promptSent,
+    });
+    return;
+  }
+
+  const reason = describeNonRatchetingReason(action);
+  logger.info(`Not ratcheting ${workspacePrPrefix} because ${reason}`, context);
+}
