+++ src/backend/services/session/service/lifecycle/session.config.service.ts
@@ -1,4 +1,4 @@
-import type { SessionConfigOption, SessionConfigSelectOption } from '@agentclientprotocol/sdk';
+import type { SessionConfigOption } from '@agentclientprotocol/sdk';
 import type { SessionPermissionPreset } from '@prisma-gen/client';
 import { createLogger } from '@/backend/services/logger.service';
 import type { AgentSessionRecord } from '@/backend/services/session/resources/agent-session.accessor';
@@ -13,11 +13,17 @@ import { userSettingsAccessor } from '@/backend/services/settings';
 import type { SessionDeltaEvent } from '@/shared/acp-protocol';
 import { type ChatBarCapabilities, EMPTY_CHAT_BAR_CAPABILITIES } from '@/shared/chat-capabilities';
 import type { SessionRepository } from './session.repository';
+import {
+  buildCapabilitiesFromConfigOptions,
+  buildCodexConfigOptionsWithModelCatalog,
+  getConfigOptionValues,
+  getSelectOptions,
+  type SessionProvider,
+} from './session-config-option-helpers';
 
 const logger = createLogger('session');
 const CODEX_MODEL_CATALOG_CACHE_TTL_MS = 30_000;
 
-type SessionProvider = 'CLAUDE' | 'CODEX';
 type SessionStartupModePreset = 'non_interactive' | 'plan';
 type CodexModelEntry = Awaited<ReturnType<typeof fetchCodexModelCatalogFromAppServer>>[number];
 type CachedCodexModelCatalog = {
@@ -253,7 +259,7 @@ export class SessionConfigService {
     if (acpHandle) {
       const modelOption = acpHandle.configOptions.find((option) => option.category === 'model');
       if (modelOption && model) {
-        const availableValues = this.getConfigOptionValues(modelOption);
+        const availableValues = getConfigOptionValues(modelOption);
         if (availableValues.length > 0 && !availableValues.includes(model)) {
           logger.debug('Skipping unsupported model for ACP session', {
             sessionId,
@@ -348,7 +354,7 @@ export class SessionConfigService {
       });
 
       if (refreshedCodexOptions.length > 0) {
-        return this.buildCapabilitiesFromConfigOptions(
+        return buildCapabilitiesFromConfigOptions(
           'CODEX',
           refreshedCodexOptions,
           cachedSnapshot?.observedModelId ?? session.model
@@ -362,7 +368,7 @@ export class SessionConfigService {
     }
 
     if (cachedSnapshot && cachedSnapshot.provider === session.provider) {
-      return this.buildCapabilitiesFromConfigOptions(
+      return buildCapabilitiesFromConfigOptions(
         session.provider,
         cachedSnapshot.configOptions,
         cachedSnapshot.observedModelId
@@ -373,7 +379,7 @@ export class SessionConfigService {
   }
 
   buildAcpChatBarCapabilities(handle: AcpProcessHandle): ChatBarCapabilities {
-    return this.buildCapabilitiesFromConfigOptions(
+    return buildCapabilitiesFromConfigOptions(
       handle.provider as SessionProvider,
       handle.configOptions
     );
@@ -431,7 +437,7 @@ export class SessionConfigService {
       return { configOptions: params.configOptions, didUpdate: false };
     }
 
-    const availableModeValues = this.getConfigOptionValues(modeOption);
+    const availableModeValues = getConfigOptionValues(modeOption);
     const targetMode = this.resolveStartupModeTarget(
       params.handle.provider as SessionProvider,
       params.startupModePreset,
@@ -561,13 +567,13 @@ export class SessionConfigService {
       return null;
     }
 
-    const availableValues = this.getConfigOptionValues(executionModeOption);
+    const availableValues = getConfigOptionValues(executionModeOption);
     const yoloByValue = this.findModeValue(availableValues, ['["never","danger-full-access"]']);
     if (yoloByValue) {
       return yoloByValue;
     }
 
-    const yoloByName = this.getSelectOptions(executionModeOption).find((option) =>
+    const yoloByName = getSelectOptions(executionModeOption).find((option) =>
       /yolo/i.test(option.name ?? '')
     );
     return yoloByName?.value ?? null;
@@ -597,7 +603,7 @@ export class SessionConfigService {
     executionModeOption: SessionConfigOption,
     permissionPreset: SessionPermissionPreset
   ): string | null {
-    const availableValues = this.getConfigOptionValues(executionModeOption);
+    const availableValues = getConfigOptionValues(executionModeOption);
     const preferredValuesByPreset: Record<SessionPermissionPreset, string[]> = {
       STRICT: [
         '["on-request","workspace-write"]',
@@ -621,7 +627,7 @@ export class SessionConfigService {
       return byValue;
     }
 
-    const byName = this.getSelectOptions(executionModeOption).find((option) => {
+    const byName = getSelectOptions(executionModeOption).find((option) => {
       const name = option.name ?? '';
       if (permissionPreset === 'STRICT') {
         return /on request/i.test(name);
@@ -699,7 +705,7 @@ export class SessionConfigService {
         return option;
       }
 
-      const allowedValues = this.getConfigOptionValues(option);
+      const allowedValues = getConfigOptionValues(option);
       if (allowedValues.length > 0 && !allowedValues.includes(value)) {
         throw new Error(
           `Unsupported value "${value}" for config option "${configId}"` +
@@ -732,7 +738,7 @@ export class SessionConfigService {
       return params.configOptions;
     }
 
-    const nextConfigOptions = this.buildCodexConfigOptionsWithModelCatalog(
+    const nextConfigOptions = buildCodexConfigOptionsWithModelCatalog(
       params.configOptions,
       modelCatalog,
       params.cachedSnapshot?.observedModelId ?? params.session.model
@@ -752,167 +758,6 @@ export class SessionConfigService {
     }
 
     return nextConfigOptions;
-  }
-
-  private buildCodexConfigOptionsWithModelCatalog(
-    existingConfigOptions: SessionConfigOption[],
-    modelCatalog: CodexModelEntry[],
-    fallbackModel?: string
-  ): SessionConfigOption[] {
-    if (modelCatalog.length === 0) {
-      return existingConfigOptions;
-    }
-
-    const nextConfigOptions = [...existingConfigOptions];
-    const resolvedModelValue = this.upsertCodexModelOption(
-      nextConfigOptions,
-      modelCatalog,
-      fallbackModel
-    );
-    if (!resolvedModelValue) {
-      return existingConfigOptions;
-    }
-
-    this.upsertCodexReasoningOption(nextConfigOptions, modelCatalog, resolvedModelValue);
-    return nextConfigOptions;
-  }
-
-  private upsertCodexModelOption(
-    configOptions: SessionConfigOption[],
-    modelCatalog: CodexModelEntry[],
-    fallbackModel?: string
-  ): string | null {
-    const modelOptionIndex = configOptions.findIndex((option) => option.category === 'model');
-    const existingModelOption =
-      modelOptionIndex >= 0 ? (configOptions[modelOptionIndex] ?? null) : null;
-    const currentModelValue =
-      this.readConfigOptionCurrentValue(existingModelOption) ??
-      this.toNonEmptyString(fallbackModel);
-    const resolvedModelValue = this.resolveCodexModelValue(currentModelValue, modelCatalog);
-    if (!resolvedModelValue) {
-      return null;
-    }
-
-    const normalizedModelOption: SessionConfigOption = {
-      id: existingModelOption?.id ?? 'model',
-      category: 'model',
-      name: existingModelOption?.name ?? 'Model',
-      type: 'select',
-      currentValue: resolvedModelValue,
-      options: modelCatalog.map((model) => ({
-        value: model.id,
-        name: model.displayName,
-        ...(model.description ? { description: model.description } : {}),
-      })),
-    };
-
-    if (modelOptionIndex >= 0) {
-      configOptions[modelOptionIndex] = normalizedModelOption;
-    } else {
-      configOptions.push(normalizedModelOption);
-    }
-
-    return resolvedModelValue;
-  }
-
-  private upsertCodexReasoningOption(
-    configOptions: SessionConfigOption[],
-    modelCatalog: CodexModelEntry[],
-    selectedModelId: string
-  ): void {
-    const reasoningOptionIndex = this.findReasoningOptionIndex(configOptions);
-    const existingReasoningOption =
-      reasoningOptionIndex >= 0 ? (configOptions[reasoningOptionIndex] ?? null) : null;
-    const selectedModel = modelCatalog.find((model) => model.id === selectedModelId);
-    const reasoningCatalog = (selectedModel?.supportedReasoningEfforts ?? []).filter(
-      (entry) => entry.reasoningEffort.trim().length > 0
-    );
-
-    if (reasoningCatalog.length === 0) {
-      if (reasoningOptionIndex >= 0) {
-        configOptions.splice(reasoningOptionIndex, 1);
-      }
-      return;
-    }
-
-    const reasoningSelectOptions = reasoningCatalog.map((entry) => ({
-      value: entry.reasoningEffort,
-      name: entry.reasoningEffort,
-      ...(entry.description ? { description: entry.description } : {}),
-    }));
-    const resolvedReasoningValue =
-      this.resolveCodexReasoningValue(
-        this.readConfigOptionCurrentValue(existingReasoningOption),
-        selectedModel?.defaultReasoningEffort,
-        reasoningSelectOptions.map((entry) => entry.value)
-      ) ?? reasoningSelectOptions[0]?.value;
-
-    if (!resolvedReasoningValue) {
-      if (reasoningOptionIndex >= 0) {
-        configOptions.splice(reasoningOptionIndex, 1);
-      }
-      return;
-    }
-
-    const normalizedReasoningOption: SessionConfigOption = {
-      id: existingReasoningOption?.id ?? 'reasoning_effort',
-      category: existingReasoningOption?.category ?? 'thought_level',
-      name: existingReasoningOption?.name ?? 'Reasoning Effort',
-      type: 'select',
-      currentValue: resolvedReasoningValue,
-      options: reasoningSelectOptions,
-    };
-
-    if (reasoningOptionIndex >= 0) {
-      configOptions[reasoningOptionIndex] = normalizedReasoningOption;
-    } else {
-      configOptions.push(normalizedReasoningOption);
-    }
-  }
-
-  private findReasoningOptionIndex(configOptions: SessionConfigOption[]): number {
-    return configOptions.findIndex(
-      (option) =>
-        option.id === 'reasoning_effort' ||
-        option.category === 'thought_level' ||
-        option.category === 'reasoning'
-    );
-  }
-
-  private resolveCodexModelValue(
-    currentModelValue: string | null,
-    modelCatalog: CodexModelEntry[]
-  ): string | null {
-    if (currentModelValue && modelCatalog.some((model) => model.id === currentModelValue)) {
-      return currentModelValue;
-    }
-    return modelCatalog.find((model) => model.isDefault)?.id ?? modelCatalog[0]?.id ?? null;
-  }
-
-  private resolveCodexReasoningValue(
-    currentReasoningValue: string | null,
-    defaultReasoningValue: string | null | undefined,
-    reasoningValues: string[]
-  ): string | null {
-    const values = new Set(reasoningValues);
-    if (currentReasoningValue && values.has(currentReasoningValue)) {
-      return currentReasoningValue;
-    }
-    if (defaultReasoningValue && values.has(defaultReasoningValue)) {
-      return defaultReasoningValue;
-    }
-    return reasoningValues[0] ?? null;
-  }
-
-  private readConfigOptionCurrentValue(option: SessionConfigOption | null): string | null {
-    if (!option) {
-      return null;
-    }
-    return this.toNonEmptyString(option.currentValue);
-  }
-
-  private toNonEmptyString(value: unknown): string | null {
-    return typeof value === 'string' && value.trim().length > 0 ? value : null;
   }
 
   private async getCodexModelCatalogFromAppServer(): Promise<CodexModelEntry[] | null> {
@@ -947,118 +792,6 @@ export class SessionConfigService {
     })();
 
     return await this.codexModelCatalogRequest;
-  }
-
-  private buildCapabilitiesFromConfigOptions(
-    provider: SessionProvider,
-    configOptions: SessionConfigOption[],
-    fallbackModel?: string
-  ): ChatBarCapabilities {
-    const modelOption = configOptions.find((option) => option.category === 'model');
-    const modeOption = configOptions.find((option) => option.category === 'mode');
-    const thoughtOption = configOptions.find(
-      (option) =>
-        option.category === 'thought_level' ||
-        option.id === 'reasoning_effort' ||
-        option.category === 'reasoning'
-    );
-    const selectedModel = modelOption?.currentValue
-      ? String(modelOption.currentValue)
-      : (fallbackModel ?? undefined);
-    const modelOptions = this.buildModelOptions(modelOption, selectedModel);
-    const isCodexProvider = provider === 'CODEX';
-    const reasoningOptions =
-      isCodexProvider && thoughtOption
-        ? this.getSelectOptions(thoughtOption).map((option) => ({
-            value: option.value,
-            label: option.name ?? option.value,
-            ...(option.description ? { description: option.description } : {}),
-          }))
-        : [];
-    const reasoningValues = new Set(reasoningOptions.map((option) => option.value));
-    const selectedReasoning =
-      isCodexProvider &&
-      thoughtOption?.currentValue &&
-      typeof thoughtOption.currentValue === 'string' &&
-      reasoningValues.has(thoughtOption.currentValue)
-        ? thoughtOption.currentValue
-        : undefined;
-    const modeDescriptors = modeOption
-      ? [
-          ...this.getConfigOptionValues(modeOption),
-          ...this.getSelectOptions(modeOption)
-            .map((entry) => entry.name ?? '')
-            .filter((value) => value.trim().length > 0),
-        ]
-      : [];
-    const planModeEnabled = modeDescriptors.some((entry) => /plan/i.test(entry));
-
-    return {
-      provider,
-      model: {
-        enabled: modelOptions.length > 0,
-        options: modelOptions,
-        ...(selectedModel ? { selected: selectedModel } : {}),
-      },
-      reasoning: {
-        enabled: reasoningOptions.length > 0,
-        options: reasoningOptions,
-        ...(selectedReasoning ? { selected: selectedReasoning } : {}),
-      },
-      thinking: {
-        enabled: !isCodexProvider && !!thoughtOption,
-      },
-      planMode: { enabled: planModeEnabled },
-      attachments: isCodexProvider
-        ? { enabled: false, kinds: [] }
-        : { enabled: true, kinds: ['image', 'text'] },
-      slashCommands: { enabled: false },
-      usageStats: { enabled: false, contextWindow: false },
-      rewind: { enabled: false },
-    };
-  }
-
-  private buildModelOptions(
-    modelOption: SessionConfigOption | undefined,
-    selectedModel: string | undefined
-  ): Array<{ value: string; label: string }> {
-    if (!modelOption) {
-      return selectedModel ? [{ value: selectedModel, label: selectedModel }] : [];
-    }
-
-    const byValue = new Map<string, string>();
-    for (const option of this.getSelectOptions(modelOption)) {
-      if (!byValue.has(option.value)) {
-        byValue.set(option.value, option.name ?? option.value);
-      }
-    }
-    if (selectedModel && !byValue.has(selectedModel)) {
-      byValue.set(selectedModel, selectedModel);
-    }
-
-    return Array.from(byValue.entries()).map(([value, label]) => ({ value, label }));
-  }
-
-  private getSelectOptions(option: SessionConfigOption): SessionConfigSelectOption[] {
-    return option.options.flatMap((entry) => {
-      if (typeof entry !== 'object' || entry === null) {
-        return [];
-      }
-      if ('value' in entry && typeof entry.value === 'string') {
-        return [entry];
-      }
-      if ('options' in entry && Array.isArray(entry.options)) {
-        return entry.options.filter(
-          (grouped): grouped is SessionConfigSelectOption =>
-            typeof grouped === 'object' && grouped !== null && typeof grouped.value === 'string'
-        );
-      }
-      return [];
-    });
-  }
-
-  private getConfigOptionValues(option: SessionConfigOption): string[] {
-    return this.getSelectOptions(option).map((entry) => entry.value);
   }
 
   private buildSnapshotPersistUpdate(
